Refact: Change todo container more componentize
