+++ YeeChanggun1902/src/components/Todo/TodoHeader.tsx
@@ -1,11 +1,8 @@
 import * as React from "react";
-import '../styles/components/TodoHeader.css';
-import {TodoModel} from "../models/TodoModel";
-import {FILTER_STATES} from "../static";
+import '../../styles/components/TodoHeader.css';
+import {TodoModel} from "../../models/TodoModel";
 
 interface CardHeaderProps{
-    filterState:FILTER_STATES,
-    changeFilter:(state:FILTER_STATES)=>void;
     addTodo:(todo:TodoModel)=>void;
 }
 
@@ -22,25 +19,12 @@ export default class TodoHeader extends React.Component<CardHeaderProps,CardHead
     }
 
     render(){
-        const {changeFilter,filterState} = this.props;
-
-        const filterDefaultClass='btn mx-1';
-        const filterElements=Object.keys(FILTER_STATES).map((state)=>
-            (<a key={state} className={filterState===FILTER_STATES[state]? `${filterDefaultClass} btn-primary`:filterDefaultClass} onClick={this.changeFilterByClick.bind(this,state)}>{state}</a>))
-
         return (
-            <div className="card-header">
+            <div className="todo-header">
                 <div className="card-title h1">TodoList</div>
                 <div className="d-flex col-12 m-2">
                     <input className="form-input mx-1" type="text" value={this.state.text} onChange={this.onTodoInputChange} placeholder="해야할 일을 입력해 보세요!"></input>
                     <button className="btn btn-primary" onClick={this.submitTodo}>제출</button>
-                </div>
-                <div className="divider"></div>
-                <div>
-                    <span className="h4 filter">FILTER: </span>
-                        {
-                            filterElements
-                        }
                 </div>
             </div>
         )
@@ -57,9 +41,5 @@ export default class TodoHeader extends React.Component<CardHeaderProps,CardHead
         this.setState({
             text:e.target.value
         })
-    }
-
-    private changeFilterByClick=(state:FILTER_STATES):void=>{
-        this.props.changeFilter(FILTER_STATES[state])
     }
 }
